+++ src/test/kotlin/de/smartsquare/socketio/emitter/JedisPublisherTest.kt
@@ -2,7 +2,6 @@ package de.smartsquare.socketio.emitter
 
 import com.redis.testcontainers.RedisContainer
 import org.amshove.kluent.shouldBeEqualTo
-import org.amshove.kluent.shouldBeTrue
 import org.amshove.kluent.shouldContain
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
@@ -10,10 +9,6 @@ import org.junit.jupiter.api.Test
 import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.junit.jupiter.Testcontainers
 import redis.clients.jedis.JedisPool
-import redis.clients.jedis.JedisPubSub
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.Executors
-import java.util.concurrent.TimeUnit
 
 @Testcontainers
 class JedisPublisherTest {
@@ -37,29 +32,11 @@ class JedisPublisherTest {
     fun `publish string message`() {
         val publisher = Emitter(JedisPublisher(pool))
 
-        val countDownLatch = CountDownLatch(1)
-
-        val listener = object : JedisPubSub() {
-            override fun onMessage(channel: String, message: String) {
-                channel shouldBeEqualTo "socket.io#/#"
-                message shouldContain "test 123"
-
-                countDownLatch.countDown()
-            }
-        }
-
-        val executor = Executors.newSingleThreadExecutor()
-        val jedis = pool.resource
-
-        try {
-            executor.submit { jedis.subscribe(listener, "socket.io#/#") }
-
+        val (channel, message) = awaitRedisMessage(redis.redisURI, "socket.io#/#") {
             publisher.broadcast("topic", "test 123")
-
-            countDownLatch.await(5, TimeUnit.SECONDS).shouldBeTrue()
-        } finally {
-            listener.unsubscribe()
-            executor.shutdown()
         }
+
+        channel shouldBeEqualTo "socket.io#/#"
+        message shouldContain "test 123"
     }
 }
